refactor: admin 강의 관련 타입 및 유틸리티 함수 정리, 비디오 썸네일 URL 처리 개선 #69
