+++ types/lecture.ts
@@ -9,6 +9,19 @@ export interface Lecture {
   challenge_id: string;
 }
 
+export interface AdminLecture {
+  id: number;
+  name: string;
+  description: string;
+  url: string;
+  created_at: string;
+  updated_at: string;
+  upload_type: number;
+  sequence: number;
+  assignment_title?: string;
+  assignment?: string;
+}
+
 export interface LectureData {
   id?: number;
   name: string;
